refactor login GUI to fix bug and set up structure
refactor GUI implementation for three login-related QDialogs to fix bugs
move all GUI design and implementation file to separated folder
backup original files for checks
add login interface test script
