--- RegisterDialog.ui
@@ -1,176 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>RegisterDialog</class>
- <widget class="QDialog" name="RegisterDialog">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>300</width>
-    <height>400</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Register</string>
-  </property>
-  <widget class="QFrame" name="frame">
-   <property name="geometry">
-    <rect>
-     <x>25</x>
-     <y>25</y>
-     <width>250</width>
-     <height>350</height>
-    </rect>
-   </property>
-   <property name="frameShape">
-    <enum>QFrame::StyledPanel</enum>
-   </property>
-   <property name="frameShadow">
-    <enum>QFrame::Raised</enum>
-   </property>
-   <widget class="QPushButton" name="Register">
-    <property name="geometry">
-     <rect>
-      <x>25</x>
-      <y>230</y>
-      <width>90</width>
-      <height>28</height>
-     </rect>
-    </property>
-    <property name="font">
-     <font>
-      <family>宋体</family>
-     </font>
-    </property>
-    <property name="text">
-     <string>注册</string>
-    </property>
-   </widget>
-   <widget class="QPushButton" name="Cancel">
-    <property name="geometry">
-     <rect>
-      <x>135</x>
-      <y>230</y>
-      <width>90</width>
-      <height>28</height>
-     </rect>
-    </property>
-    <property name="font">
-     <font>
-      <family>宋体</family>
-     </font>
-    </property>
-    <property name="text">
-     <string>取消</string>
-    </property>
-   </widget>
-   <widget class="QLabel" name="Title">
-    <property name="geometry">
-     <rect>
-      <x>65</x>
-      <y>45</y>
-      <width>120</width>
-      <height>30</height>
-     </rect>
-    </property>
-    <property name="font">
-     <font>
-      <family>等线</family>
-      <pointsize>12</pointsize>
-     </font>
-    </property>
-    <property name="layoutDirection">
-     <enum>Qt::LeftToRight</enum>
-    </property>
-    <property name="text">
-     <string>XXX数据库</string>
-    </property>
-    <property name="textFormat">
-     <enum>Qt::AutoText</enum>
-    </property>
-    <property name="alignment">
-     <set>Qt::AlignCenter</set>
-    </property>
-   </widget>
-   <widget class="QLineEdit" name="PasswordAgain">
-    <property name="geometry">
-     <rect>
-      <x>45</x>
-      <y>180</y>
-      <width>160</width>
-      <height>30</height>
-     </rect>
-    </property>
-    <property name="font">
-     <font>
-      <family>宋体</family>
-     </font>
-    </property>
-    <property name="text">
-     <string>确认密码</string>
-    </property>
-   </widget>
-   <widget class="QLineEdit" name="Password">
-    <property name="geometry">
-     <rect>
-      <x>45</x>
-      <y>140</y>
-      <width>160</width>
-      <height>30</height>
-     </rect>
-    </property>
-    <property name="font">
-     <font>
-      <family>宋体</family>
-     </font>
-    </property>
-    <property name="autoFillBackground">
-     <bool>false</bool>
-    </property>
-    <property name="text">
-     <string>密码</string>
-    </property>
-   </widget>
-   <widget class="QLineEdit" name="UserName">
-    <property name="geometry">
-     <rect>
-      <x>45</x>
-      <y>100</y>
-      <width>160</width>
-      <height>30</height>
-     </rect>
-    </property>
-    <property name="font">
-     <font>
-      <family>宋体</family>
-     </font>
-    </property>
-    <property name="autoFillBackground">
-     <bool>false</bool>
-    </property>
-    <property name="text">
-     <string>用户名</string>
-    </property>
-   </widget>
-  </widget>
- </widget>
- <resources/>
- <connections>
-  <connection>
-   <sender>Register</sender>
-   <signal>clicked()</signal>
-   <receiver>RegisterDialog</receiver>
-   <slot>accept()</slot>
-   <hints>
-    <hint type="sourcelabel">
-     <x>103</x>
-     <y>272</y>
-    </hint>
-    <hint type="destinationlabel">
-     <x>132</x>
-     <y>393</y>
-    </hint>
-   </hints>
-  </connection>
- </connections>
-</ui>
